SSR: trust guessedBreakpoint during hydratation
As explained in #29, when the breapoint is wrongly guessed on the server, there will be an HTML mismatch during `ReactDom.hydrate()`. When using a CSS-in-JSS technique, this can lead to a completely broken page due to wrong CSS classnames being used.

This PR simply solves this issue by letting the developer set the `guessedBreakpoint` on the client as well. Thanks to this, there will no mismatch during hydratation. Once the app is hydrated, it will be rerenderd using the real breakpoint (during `componentDidMount()` ).

There will be a "blinking" effect on load when switching between the guesss breakpoint and the real one (as with current version) but now, there is no mismatch so no risk of unfortunate side-effects anymore.